WIP: Start adding lib-param-bn bindings to new project format.
