+++ src/bindings/lib_param_bn/_impl_variable_id.rs
@@ -1,0 +1,26 @@
+use biodivine_lib_param_bn::VariableId;
+use pyo3::prelude::*;
+use crate::bindings::lib_param_bn::PyVariableId;
+
+impl From<PyVariableId> for VariableId {
+    fn from(value: PyVariableId) -> Self {
+        value.0
+    }
+}
+
+impl From<VariableId> for PyVariableId {
+    fn from(value: VariableId) -> Self {
+        PyVariableId(value)
+    }
+}
+
+#[pymethods]
+impl PyVariableId {
+    fn __str__(&self) -> PyResult<String> {
+        Ok(format!("BnVariableId({})", self.0))
+    }
+
+    fn __repr__(&self) -> PyResult<String> {
+        self.__str__()
+    }
+}
