+++ src/bindings/mod.rs
@@ -1,3 +1,6 @@
 
 /// Wrappers for the `biodivine-lib-bdd` library.
 pub mod lib_bdd;
+
+/// Wrappers for the `biodivine-lib-param-bn` library.
+pub mod lib_param_bn;
